Use size_t over int for vlen.
git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@10599 221aa14e-8319-0410-a670-987f0aec2ac5

+++ gnuradio-core/src/lib/general/gr_float_to_complex.cc
@@ -28,12 +28,12 @@
 #include <gr_io_signature.h>
 
 gr_float_to_complex_sptr
-gr_make_float_to_complex (int vlen)
+gr_make_float_to_complex (size_t vlen)
 {
   return gr_float_to_complex_sptr (new gr_float_to_complex (vlen));
 }
 
-gr_float_to_complex::gr_float_to_complex (int vlen)
+gr_float_to_complex::gr_float_to_complex (size_t vlen)
   : gr_sync_block ("gr_float_to_complex",
 		   gr_make_io_signature (1, 2, sizeof (float) *  vlen),
 		   gr_make_io_signature (1, 1, sizeof (gr_complex) * vlen)),
